@@ -23,7 +23,6 @@ import fr.olympa.api.utils.Prefix;
 import fr.olympa.core.spigot.OlympaCore;
 import fr.olympa.zta.OlympaZTA;
 import fr.olympa.zta.ZTAPermissions;
-import fr.olympa.zta.glass.GlassSmashFlag;
 import fr.olympa.zta.mobs.MobSpawning;
 import fr.olympa.zta.mobs.MobSpawning.SpawnType;
 import fr.olympa.zta.utils.DynmapLink;
@@ -55,8 +54,7 @@ public class HubManager implements Listener {
 				new DynmapLink.DynmapHideFlag(),
 				new FoodFlag(true, true, false),
 				new GunFlag(true, false),
-				new DamageFlag(true).setMessages("§e§lBienvenue au Hub !", null, ChatMessageType.ACTION_BAR),
-				new GlassSmashFlag(true));
+				new DamageFlag(true).setMessages("§e§lBienvenue au Hub !", null, ChatMessageType.ACTION_BAR));
 	}
 
 	public void addSpawnRegion(SpawnType region) {
